pyproject.toml
[tool.poetry]
name = "candlestick_aggregator"
version = "0.1.0"
description = ""
authors = ["Gabriel Pires <gabrielhpires@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
mysql-connector = "^2.2.9"
retry = "^0.9.2"
requests = "^2.25.1"
freezegun = "^1.1.0"
pandas = "^1.2.4"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-mock = "^3.6.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
